+++ api/Pipfile
@@ -1,0 +1,21 @@
+[[source]]
+name = "pypi"
+url = "https://pypi.org/simple"
+verify_ssl = true
+
+[dev-packages]
+pysqlite3 = "*"
+
+[packages]
+flask = "*"
+lxml = "*"
+eospy = "*"
+Flask = "*"
+peewee = "*"
+gunicorn = "*"
+supervisor = "*"
+libeospy = "*"
+psycopg2-binary = "*"
+
+[requires]
+python_version = "3.7"
