+++ api/index.py
@@ -41,6 +41,11 @@ def test_net_is_working(urls):
 
 
 app = Flask(__name__)
+app.url_map.strict_slashes = False
+
+@app.route('/')
+def hi():
+    return '🐛'
 
 @app.route('/getBPs')
 def bpinfos():
@@ -84,6 +89,8 @@ def jsoninfo():
 	for proposal in proposals:
 		proposal_item = proposals[proposal]
 		vote_key = proposal_item['stats']['votes']
+		stake_key = proposal_item['stats']['staked']
+		staked_total = float(int(proposal_item['stats']['staked']['total'])/10000) if 'total' in stake_key else 0
 		vote_total = proposal_item['stats']['votes']['total'] if  'total' in vote_key else 0
 		vote_yes = proposal_item['stats']['votes']['1'] if  '1' in vote_key else 0
 		vote_no = proposal_item['stats']['votes']['0'] if '0' in vote_key else 0
@@ -94,6 +101,8 @@ def jsoninfo():
 		_reviewed_by_BET_date =  None
 		_approved_by_BPs =  0
 		_approved_by_BPs_date =  None
+		_finish =  0
+		_finish_date =  None
 		# check if the proposal exists
 		try:
 			propos = Proposal.select().where(Proposal.name == proposal).count()
@@ -101,7 +110,7 @@ def jsoninfo():
 			# the proposal exists
 			if propos > 0:
 				# check if the proposal is passed
-				if proposal_base_condition_ckeck(BP_TOTAL_VOTES, vote_total, vote_yes, vote_no):
+				if proposal_base_condition_ckeck(BP_TOTAL_VOTES, staked_total, vote_yes, vote_no):
 					propos = Proposal.get(Proposal.name == proposal)
 					meet = propos.meet_conditions_days
 					# abv = propos.approved_by_vote 
@@ -112,32 +121,34 @@ def jsoninfo():
 					nrow = (Proposal.update(
 						json_info = proposal_item
 						, vote_total = vote_total
+						, staked_total = staked_total
+						, vote_bp_total = BP_TOTAL_VOTES
 						, vote_yes = vote_yes
 						, vote_no = vote_no
 						, meet_conditions_days = int(meet) + 1
 						, approved_by_vote = _approved_by_vote
 						, approved_by_vote_date = _approved_by_vote_date
-						# , approved_by_BET = _approved_by_BET
-						# , reviewed_by_BET_date = _reviewed_by_BET_date
-						# , approved_by_BPs = _approved_by_BPs
-						# , approved_by_BPs_date = _approved_by_BPs_date
 						).where(Proposal.name == proposal).execute())
 				else:
 					nrow = (Proposal.update(
 						json_info = proposal_item
 						, vote_total = vote_total
+						, staked_total = staked_total
+						, vote_bp_total = BP_TOTAL_VOTES
 						, vote_yes = vote_yes
 						, vote_no = vote_no
 						, meet_conditions_days = 0
 						, approved_by_vote = 0
 						, approved_by_vote_date = None
-						# , approved_by_BET = _approved_by_BET
-						# , reviewed_by_BET_date = _reviewed_by_BET_date
-						# , approved_by_BPs = _approved_by_BPs
-						# , approved_by_BPs_date = _approved_by_BPs_date
+						, approved_by_BET = 0
+						, reviewed_by_BET_date = None
+						, approved_by_BPs = 0
+						, approved_by_BPs_date = None
+						, finish = 0
+						, finish_date = None
 						).where(Proposal.name == proposal).execute())
 			else:
-				if proposal_base_condition_ckeck(BP_TOTAL_VOTES, vote_total, vote_yes, vote_no):
+				if proposal_base_condition_ckeck(BP_TOTAL_VOTES, staked_total, vote_yes, vote_no):
 					_approved_by_vote = 1
 					_approved_by_vote_date = datetime.now()
 				else:
@@ -145,6 +156,8 @@ def jsoninfo():
 				new_proposal = Proposal(name = proposal
 				, json_info = proposal_item
 				, vote_total = vote_total
+				, staked_total = staked_total
+				, vote_bp_total = BP_TOTAL_VOTES
 				, vote_yes = vote_yes
 				, vote_no = vote_no
 				, meet_conditions_days = _meet_conditions_days
@@ -153,12 +166,15 @@ def jsoninfo():
 				, approved_by_BET = _approved_by_BET
 				, reviewed_by_BET_date = _reviewed_by_BET_date
 				, approved_by_BPs = _approved_by_BPs
-				, approved_by_BPs_date = _approved_by_BPs_date)
+				, approved_by_BPs_date = _approved_by_BPs_date
+				, finish = _finish
+				, finish_date = _finish_date)
 				new_proposal.save()  
 		except Exception as err:
 			logger.error(err)
 		logger.info("proposal name: " + proposal)
 		logger.info("vote total: " + str(vote_total))
+		logger.info("staked total: " + str(staked_total))
 		logger.info("vote yes: " + str(vote_yes))
 		logger.info("vote no: " + str(vote_no))
 		logger.info("___________________")
@@ -193,6 +209,8 @@ def proposals():
 				proposals[proposal]['reviewed_by_BET_date'] = str(propos.reviewed_by_BET_date)
 				proposals[proposal]['approved_by_BPs'] = propos.approved_by_BPs
 				proposals[proposal]['approved_by_BPs_date'] = str(propos.approved_by_BPs_date)
+				proposals[proposal]['finish'] = propos.finish
+				proposals[proposal]['finish_date'] = str(propos.finish_date)
 			else:
 				proposals[proposal]['meet_conditions_days'] = 0
 				proposals[proposal]['approved_by_vote'] = 0
@@ -201,6 +219,8 @@ def proposals():
 				proposals[proposal]['reviewed_by_BET_date'] = ""
 				proposals[proposal]['approved_by_BPs'] = 0
 				proposals[proposal]['approved_by_BPs_date'] = ""
+				proposals[proposal]['finish'] = ""
+				proposals[proposal]['finish_date'] = 0
 		except Exception as err:
 				logger.error(err)
 	resp = flask.Response(json.dumps(proposals), mimetype='application/json')
@@ -231,6 +251,8 @@ def proposal(proposal_name):
 			proposals[proposal_name]['reviewed_by_BET_date'] = str(propos.reviewed_by_BET_date)
 			proposals[proposal_name]['approved_by_BPs'] = propos.approved_by_BPs
 			proposals[proposal_name]['approved_by_BPs_date'] = str(propos.approved_by_BPs_date)
+			proposals[proposal_name]['finish'] = propos.finish
+			proposals[proposal_name]['finish_date'] = str(propos.finish_date)
 		else:
 			proposals[proposal_name]['meet_conditions_days'] = 0
 			proposals[proposal_name]['approved_by_vote'] = 0
@@ -239,8 +261,73 @@ def proposal(proposal_name):
 			proposals[proposal_name]['reviewed_by_BET_date'] = ""
 			proposals[proposal_name]['approved_by_BPs'] = 0
 			proposals[proposal_name]['approved_by_BPs_date'] = ""
+			proposals[proposal_name]['finish'] = ""
+			proposals[proposal_name]['finish_date'] = 0
 	except Exception as err:
 			logger.error(err)
 	resp = flask.Response(json.dumps(proposals[proposal_name]), mimetype='application/json')
 	resp.headers['Access-Control-Allow-Origin'] = '*'
 	return resp
+
+
+@app.route('/review/<proposal_name>')
+def review(proposal_name):
+	logger = reactive_log('review')
+	try:
+		propos = Proposal.select().where(Proposal.name == proposal_name).count()
+		logger.info("result"+str(proposal_name))
+		#the proposal exists
+		if propos > 0:
+			propos2 = Proposal.get(Proposal.name == proposal_name)
+			print(propos2)
+			if propos2.approved_by_vote == 1:
+				nrow=(Proposal.update(approved_by_BET = 1
+					,reviewed_by_BET_date = datetime.now()
+					).where(Proposal.name == proposal_name).execute())
+				resp = flask.Response(json.dumps({"result":"ok"}), mimetype='application/json')
+				resp.headers['Access-Control-Allow-Origin'] = '*'
+				return resp
+			else:
+				resp = flask.Response(json.dumps({"result":"not approved by vote"}), mimetype='application/json')
+				resp.headers['Access-Control-Allow-Origin'] = '*'
+				return resp
+		else:
+			resp = flask.Response(json.dumps({"result":"not exited"}), mimetype='application/json')
+			resp.headers['Access-Control-Allow-Origin'] = '*'
+			return resp
+	except Exception as err:
+			logger.error(err)
+			resp = flask.Response(json.dumps({"result":"ERROR"}), mimetype='application/json')
+			resp.headers['Access-Control-Allow-Origin'] = '*'
+			return resp
+
+@app.route('/finish/<proposal_name>')
+def finish(proposal_name):
+	logger = reactive_log('review')
+	try:
+		propos = Proposal.select().where(Proposal.name == proposal_name).count()
+		logger.info("result"+str(propos))
+		# the proposal exists
+		if propos > 0:
+			propos2 = Proposal.get(Proposal.name == proposal_name)
+			if propos2.approved_by_vote == 1 and propos2.approved_by_BET == 1:
+				nrow = (Proposal.update(
+							finish = 1
+							, finish_date = datetime.now()
+							).where(Proposal.name == proposal_name).execute())
+				resp = flask.Response(json.dumps({"result":"ok"}), mimetype='application/json')
+				resp.headers['Access-Control-Allow-Origin'] = '*'
+				return resp
+			else:
+				resp = flask.Response(json.dumps({"result":"not approved by vote and not review"}), mimetype='application/json')
+				resp.headers['Access-Control-Allow-Origin'] = '*'
+				return resp
+		else:
+			resp = flask.Response(json.dumps({"result":"not exited"}), mimetype='application/json')
+			resp.headers['Access-Control-Allow-Origin'] = '*'
+			return resp
+	except Exception as err:
+			logger.error(err)
+			resp = flask.Response(json.dumps({"result":"ERROR"}), mimetype='application/json')
+			resp.headers['Access-Control-Allow-Origin'] = '*'
+			return resp
