+++ api/now.json
@@ -1,0 +1,11 @@
+{
+  "version": 2,
+  "name": "boscore-referendum-api",
+  "builds": [
+    { "src": "**/index.py", "use": "@now/python", "config": { "maxLambdaSize": "20mb" } }
+  ],
+  "routes": [{ "src": "(.*)", "dest": "index.py" }],
+  "env": {
+    "DB_PWD": "@my-db-pw"
+  }
+}
