--- backend-api/.gitignore
@@ -1,5 +1,0 @@
-log/
-__pycache__/
-.history/
-.vscode/
-*.db
