--- backend-api/init_run_dev.sh
@@ -1,3 +1,0 @@
-export FLASK_APP=api.py
-export FLASK_ENV=development
-python3 -m flask run
