--- backend-api/package.txt
@@ -1,9 +1,0 @@
-lxml
-eospy
-Flask
-peewee
-gunicorn
-supervisor
-
-UnboundLocalError: local variable 'proposals' referenced before assignment
-先初始化数据库然后 /getJson
