--- backend-api/sumary.sh
@@ -1,1 +1,0 @@
-curl localhost:5000/getJson
